+++ modules/apps/commerce/commerce-order-web/src/main/java/com/liferay/commerce/order/web/internal/frontend/taglib/form/navigator/CommerceOrderDetailsFormNavigatorCategory.java
@@ -14,7 +14,7 @@
 
 package com.liferay.commerce.order.web.internal.frontend.taglib.form.navigator;
 
-import com.liferay.commerce.order.web.internal.servlet.taglib.ui.constants.CommerceOrderFormNavigatorConstants;
+import com.liferay.commerce.order.web.internal.constants.CommerceOrderFormNavigatorConstants;
 import com.liferay.frontend.taglib.form.navigator.FormNavigatorCategory;
 import com.liferay.portal.kernel.language.Language;
